+++ src/main/java/com/martinheywang/model/devices/behaviours/Seller_.java
@@ -27,8 +27,10 @@ public class Seller_ extends Behaviour {
 			else {
 				if (!resATraiter.getRessource().equals(BaseResources.NONE)) {
 					controller.addMoney(
-							resATraiter.getRessource().getPrice() - Device
-									.getElectricity());
+							BigInteger.valueOf(
+									resATraiter.getRessource().getPrice()
+											- Device
+													.getElectricity()));
 					device.activate();
 				}
 			}
